meta: Refactor plant.ts to modular sub-modules under utils/plants/
- Split frontend/src/utils/plant.ts (3,005 lines) into 6 dedicated modular sub-files:
  - types.ts: Shared interfaces (PlantGenomeData, FloralGenome, structures).
  - helpers.ts: Seeded random and HSL color conversion functions.
  - lsystem.ts: L-system string production and geometry parsing.
  - textures.ts: Fractal canvas texture generators (mandelbrot, claude, antigravity, gpt).
  - renderers.ts: Custom renderers for GPT Codex and Sonnet 4.5 plants.
  - nectar.ts: Nectar item rendering and the 10+ floral fractal curves.
- plant.ts remains as the main entry point (retaining full public API compatibility).
- Reduced plant.ts size from 3,005 lines to 503 lines.
- All frontend tests, TypeScript builds, and Python smoke/agent gates pass.
